Hotfix: Settings context (Into 0.3.0) (#77)
