+++ src/hooks.tsx
@@ -1,7 +1,5 @@
 import { useRouter } from 'next/navigation';
 import { useState, useEffect } from 'react';
-import { getSettings } from '@/api/users/settings';
-import { SettingsContextInterface, defaultSettings } from './context';
 
 export const useLogin = () => {
   const router = useRouter();
@@ -24,32 +22,6 @@ export const useLogin = () => {
     uid,
     email,
   };
-};
-
-export const useSettings = async (uid: string) => {
-  const router = useRouter();
-
-  const [settings, setUserSettings] = useState<SettingsContextInterface>(defaultSettings);
-
-  async function fetchSettings() {
-    const settings = await getSettings(uid);
-    setUserSettings(settings);
-    return settings;
-  }
-
-  useEffect(() => {
-    if (uid == 'ERROR') {
-      localStorage.removeItem('uid');
-      localStorage.removeItem('email');
-      router.push('/');
-    } else {
-      if (uid != '') {
-        fetchSettings();
-      }
-    }
-  }, [uid]);
-
-  return { settings, fetchSettings };
 };
 
 export default useLogin;
